fix(54): resolve 426 build errors from Phase 54 parallel development
Base class contract drift from 11 parallel plans caused CS0534, CS0115,
CS0246, CS1503, CS0104, CS0108, CS8602, CS0101 errors across 16 projects.
All fixed mechanically: missing overrides added, type conflicts resolved,
null safety applied, ambiguous references qualified.

+++ Plugins/DataWarehouse.Plugins.UltimateConsensus/IRaftStrategy.cs
@@ -549,7 +549,7 @@ public sealed class PbftStrategy : IRaftStrategy
 
     private void ApplyCommittedEntries(PbftGroupState state)
     {
-        while (state.CommittedEntries.ContainsKey(state.LastApplied + 1))
+        while (state.CommittedEntries.ContainsKey((int)(state.LastApplied + 1)))
         {
             state.LastApplied++;
             // Entry applied to state machine
